+++ reactjs/src/components/PageRating.js
@@ -1,8 +1,10 @@
 import React from 'react'
-import AnyCommentComponent from "./AnyCommentComponent";
 import Icon from './Icon'
-import {faStar, faStarHalfAlt} from '@fortawesome/free-solid-svg-icons'
+import Rating from 'react-rating'
 import {toast} from 'react-toastify'
+import AnyCommentComponent from "./AnyCommentComponent";
+import {faStar} from '@fortawesome/free-solid-svg-icons'
+import {faStar as faStarEmpty} from '@fortawesome/free-regular-svg-icons'
 
 /**
  * Component used to display rating.
@@ -10,7 +12,7 @@ import {toast} from 'react-toastify'
 class PageRating extends AnyCommentComponent {
 
     constructor(props) {
-        super(props);
+        super();
 
         this.state = {
             value: props.settings.rating.value,
@@ -22,13 +24,10 @@ class PageRating extends AnyCommentComponent {
     /**
      * Action to set rating.
      *
-     * @param e
      * @param rating
      * @returns {*}
      */
-    rate = (e, rating) => {
-        e.preventDefault();
-
+    onRate = (rating) => {
         const settings = this.getSettings(),
             self = this;
 
@@ -66,39 +65,6 @@ class PageRating extends AnyCommentComponent {
     };
 
     /**
-     * Render stars.
-     *
-     * @returns {Array}
-     */
-    renderStars() {
-        let stars = [];
-
-        let {value} = this.state;
-
-        const matches = value > 0 ? value.match(/([1-5])\.([0-9])/m) : [];
-
-        for (let i = 5; i >= 1; i--) {
-
-            const isActive = i <= value,
-                activeClass = isActive ? " anycomment-rating__stars-item-active" : '';
-
-            let icon = faStar;
-
-            if (matches !== [] && parseInt(matches[1]) === i && matches[2] >= 5) {
-                icon = faStarHalfAlt;
-            }
-
-            const item = <span
-                className={"anycomment anycomment-rating__stars-item" + activeClass}
-                onClick={(e) => this.rate(e, i)}><Icon size={24} icon={icon}/></span>;
-
-            stars.push(item);
-        }
-
-        return stars;
-    }
-
-    /**
      * Render component.
      *
      * @returns {*}
@@ -117,7 +83,16 @@ class PageRating extends AnyCommentComponent {
                  className="anycomment anycomment-rating">
                 <div
                     className={"anycomment anycomment-rating__stars" + (hasRated ? " anycomment-rating__stars-readonly" : '')}>
-                    {this.renderStars()}
+                    <Rating
+                        start={0}
+                        stop={5}
+                        step={1}
+                        initialRating={value}
+                        readonly={hasRated}
+                        emptySymbol={<Icon size="22px" icon={faStarEmpty}/>}
+                        fullSymbol={<Icon class="anycomment anycomment-rating__stars-active" color="#eeba64" size="22px" icon={faStar}/>}
+                        onClick={this.onRate}
+                    />
                 </div>
                 <div className="anycomment anycomment-rating__count"
                      itemProp="aggregateRating"